leave log files if not verbose

+++ mt-kahypar/partition/refinement/spectral/solvers/julia/tree_partition/metis.jl
@@ -119,6 +119,8 @@ function metis(graph_file::String,
 
     run(`$sh_file`, wait=true)
 
-    rm_cmd = "rm -r $log_file $graph_file $sh_file"
-    run(`sh -c $rm_cmd`, wait=true)
+    if config_verbose
+        rm_cmd = "rm -r $log_file $graph_file $sh_file"
+        run(`sh -c $rm_cmd`, wait=true)
+    end
 end
