+++ mt-kahypar/partition/refinement/spectral/solvers/julia/tree_partition/tree_partition.jl
@@ -536,7 +536,9 @@ function tree_partition(adj::SparseMatrixCSC,
             end
         end
     end
-    rm_cmd = "rm -r " * config_tmpDir * "/*.part." * string(config_k)
-    run(`sh -c $rm_cmd`, wait=true)
+    if config_verbose
+        rm_cmd = "rm -r " * config_tmpDir * "/*.part." * string(config_k)
+        run(`sh -c $rm_cmd`, wait=true)
+    end
     return partitions
 end
